refactor(mcp): extract Server Tools and trim contract/facade docs
